Fix dialyzer warning in simple_bridge_websocket:apply_mask/2.
Previously, dialyzer complained:
  simple_bridge_websocket.erl:441: The pattern
  <<<Mask:32/integer-unit:1>>, Data> can never match the type
  <non_neg_integer(),bitstring()>

I believe that the apply_mask(binary(), term()) head is unnecessary,
but instead of removing it I made a dummy call to silence dialyzer's
warning.  It's also not clear to me why the execution time of
apply_mask/3 is being timed, since the time is being discarded.