Add `simple_bridge:host/1` callback
The `host` value for a given request may depend on multiple values:
 - the `Host` HTTP header
 - the `X-Forwarded-For` HTTP header
 - the absolute URI for the request
 - the `Host` property of the request set by the HTTP server

When applications rely on getting this value diretly through the HTTP
headers they may run into issues with this being undefined or wrong.
Possible scenarios may include:
 - pre-HTTP/1.1 requests without an explicit `Host` HTTP header
 - TLS requests when the underlying web server doesn't set a header
   from the initial `:authority` parameter.
 - proxied requests with an `X-Forwarded-For` HTTP header

In this patch we're adding a new `host/1` callback so that the underlying
implementation can provide a proper value for the request's host, whether
or not it's set as an HTTP header, and whether or not a proxy might be
providing more accurate information.